Step 8: Move backtrace CalleeSavedRegs and naked dispatch to arch
Move CalleeSavedRegs struct with its naked dispatch function to
arch/src/riscv64/backtrace.rs (RISC-V specific assembly). Add stub
for non-RISC-V targets. Add sys::memory::read_usize_from_address
to encapsulate stack pointer reads.

Kernel's backtrace.rs no longer has #![allow(unsafe_code)] on the
main module (only the test submodule retains it for the unwinding
crate's callback).
