+++ arch/src/stub/mod.rs
@@ -1,3 +1,4 @@
+pub mod backtrace;
 pub mod cpu;
 pub mod linker_symbols;
 pub mod sbi;
