+++ kernel/src/debugging/backtrace.rs
@@ -1,7 +1,5 @@
-#![allow(unsafe_code)]
 use super::eh_frame_parser;
 use crate::{
-    assert::static_assert_size,
     cpu::KERNEL_STACK_SIZE,
     debugging::{
         self,
@@ -13,11 +11,7 @@ use crate::{
     memory::{address::VirtAddr, linker_information::LinkerInformation},
 };
 use alloc::vec::Vec;
-// Needed for the native backtrace impl for debugging purposes
-// use core::ffi::c_void;
-// use unwinding::abi::{
-//     UnwindContext, UnwindReasonCode, _Unwind_Backtrace, _Unwind_GetIP, with_context,
-// };
+use arch::backtrace::CalleeSavedRegs;
 
 #[allow(dead_code)]
 #[derive(Debug)]
@@ -31,10 +25,6 @@ fn is_in_text_segment(address: usize) -> bool {
     LinkerInformation::text_range().contains(&VirtAddr::new(address))
 }
 
-/// We keep the already parsed information in a Vec
-/// even though we might not even need to produce a backtrace
-/// But we want to avoid heap allocation while backtracing
-/// in case of memory corruption.
 struct Backtrace<'a> {
     fdes: Vec<eh_frame_parser::ParsedFDE<'a>>,
 }
@@ -87,10 +77,6 @@ impl<'a> Backtrace<'a> {
             return Err(BacktraceNextError::RaOutsideText(ra));
         }
 
-        // RA points to the next instruction. Move it back one byte such
-        // that it points into the previous instruction.
-        // This case must be handled different as soon as we have
-        // signal trampolines.
         let fde = self
             .find(ra - 1)
             .ok_or(BacktraceNextError::CouldNotGetFde(ra))?;
@@ -114,10 +100,8 @@ impl<'a> Backtrace<'a> {
                     continue;
                 }
                 RegisterRule::Offset(offset) => {
-                    let ptr = crate::klibc::util::wrapping_add_signed(cfa, *offset) as *const usize;
-                    // SAFETY: ptr is CFA + offset, which points to the saved
-                    // register value on the stack frame.
-                    unsafe { ptr.read() }
+                    let addr = crate::klibc::util::wrapping_add_signed(cfa, *offset);
+                    sys::memory::read_usize_from_address(addr)
                 }
             };
             new_regs[reg_index] = value;
@@ -126,219 +110,6 @@ impl<'a> Backtrace<'a> {
         *regs = new_regs;
 
         Ok(ra)
-    }
-}
-
-// We leave that here for debugging purposes
-// I'm not entirely sure if my own backtrace implementation
-// is fault free. But we will see that in the future.
-// After multiple months of implementing this I'm done and want to move forward
-// to something else.
-// fn print_native() {
-//     #[derive(Default)]
-//     struct CallbackData {
-//         counter: usize,
-//     }
-
-//     extern "C" fn callback(unwind_ctx: &UnwindContext<'_>, arg: *mut c_void) -> UnwindReasonCode {
-//         let data = unsafe { &mut *(arg as *mut CallbackData) };
-//         data.counter += 1;
-//         info!("{}: {:#x}", data.counter, _Unwind_GetIP(unwind_ctx));
-//         UnwindReasonCode::NO_REASON
-//     }
-
-//     let mut data = CallbackData::default();
-
-//     _Unwind_Backtrace(callback, &mut data as *mut _ as _);
-// }
-
-/// You ask where I got the registers from? This is a good question.
-/// I just looked what registers were mentioned in the eh_frame and added those.
-/// Maybe there will be more in the future, then we have to add them.
-/// I tried to generate the following code via a macro. However this is not possible,
-/// because they won't allow to concatenate x$num_reg as a identifier and I need the
-/// literal number to access it via an index.
-#[derive(Debug, Clone, Default)]
-struct CalleeSavedRegs {
-    x1: usize,
-    x2: usize,
-    x8: usize,
-    x9: usize,
-    x18: usize,
-    x19: usize,
-    x20: usize,
-    x21: usize,
-    x22: usize,
-    x23: usize,
-    x24: usize,
-    x25: usize,
-    x26: usize,
-    x27: usize,
-}
-
-impl core::fmt::Display for CalleeSavedRegs {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        macro_rules! print_reg {
-            ($reg:ident) => {
-                writeln!(f, "{}: {:#x}", stringify!($reg), self.$reg)?
-            };
-        }
-
-        print_reg!(x1);
-        print_reg!(x2);
-        print_reg!(x8);
-        print_reg!(x9);
-        print_reg!(x18);
-        print_reg!(x19);
-        print_reg!(x20);
-        print_reg!(x21);
-        print_reg!(x22);
-        print_reg!(x23);
-        print_reg!(x24);
-        print_reg!(x25);
-        print_reg!(x26);
-        print_reg!(x27);
-
-        Ok(())
-    }
-}
-
-impl core::ops::Index<usize> for CalleeSavedRegs {
-    type Output = usize;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        match index {
-            1 => &self.x1,
-            2 => &self.x2,
-            8 => &self.x8,
-            9 => &self.x9,
-            18 => &self.x18,
-            19 => &self.x19,
-            20 => &self.x20,
-            21 => &self.x21,
-            22 => &self.x22,
-            23 => &self.x23,
-            24 => &self.x24,
-            25 => &self.x25,
-            26 => &self.x26,
-            27 => &self.x27,
-            _ => panic!("Invalid index"),
-        }
-    }
-}
-
-impl core::ops::IndexMut<usize> for CalleeSavedRegs {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        match index {
-            1 => &mut self.x1,
-            2 => &mut self.x2,
-            8 => &mut self.x8,
-            9 => &mut self.x9,
-            18 => &mut self.x18,
-            19 => &mut self.x19,
-            20 => &mut self.x20,
-            21 => &mut self.x21,
-            22 => &mut self.x22,
-            23 => &mut self.x23,
-            24 => &mut self.x24,
-            25 => &mut self.x25,
-            26 => &mut self.x26,
-            27 => &mut self.x27,
-            _ => panic!("Invalid index"),
-        }
-    }
-}
-
-// This value is referenced in the assembly of extern "C-unwind" fn dispatch
-static_assert_size!(CalleeSavedRegs, 0x70);
-
-impl CalleeSavedRegs {
-    fn ra(&self) -> usize {
-        self.x1
-    }
-
-    fn set_ra(&mut self, value: usize) {
-        self.x1 = value;
-    }
-
-    fn sp(&self) -> usize {
-        self.x2
-    }
-
-    fn set_sp(&mut self, value: usize) {
-        self.x2 = value;
-    }
-
-    fn with_context<F: FnMut(&mut CalleeSavedRegs)>(f: F) {
-        // Inspired by the unwinder crate
-        // https://github.com/nbdd0121/unwinding/
-
-        // We cannot call a closure directly from assembly
-        // because we're missing some compiler magic.
-        // Convert the closure to a fn pointer by having a
-        // intermediate function closure_to_fn_pointer.
-
-        // Not the prettiest code but very cool and also
-        // very convenient for the caller side.
-
-        #[repr(C)]
-        struct ClosureWrapper<F: FnMut(&mut CalleeSavedRegs)>(F);
-
-        let mut closure = ClosureWrapper(f);
-
-        dispatch(
-            &mut CalleeSavedRegs::default(),
-            &mut closure,
-            closure_to_fn_pointer,
-        );
-
-        extern "C" fn closure_to_fn_pointer<F: FnMut(&mut CalleeSavedRegs)>(
-            regs: &mut CalleeSavedRegs,
-            f_data: &mut ClosureWrapper<F>,
-        ) {
-            (f_data.0)(regs);
-        }
-
-        // SAFETY: Naked function that captures callee-saved registers (s0-s11,
-        // ra, sp) into the CalleeSavedRegs struct, then calls the closure.
-        // No prologue is generated so we get the true register state.
-        #[unsafe(naked)]
-        extern "C-unwind" fn dispatch<F: FnMut(&mut CalleeSavedRegs)>(
-            regs: &mut CalleeSavedRegs,
-            f_data: &mut ClosureWrapper<F>,
-            f: extern "C" fn(&mut CalleeSavedRegs, &mut ClosureWrapper<F>),
-        ) {
-            core::arch::naked_asm!(
-                "
-                     # regs is in a0
-                     # f to call in a2
-                     sd x1, 0x00(a0)   
-                     sd x2, 0x08(a0)
-                     sd x8, 0x10(a0)
-                     sd x9, 0x18(a0)
-                     sd x18, 0x20(a0)
-                     sd x19, 0x28(a0)
-                     sd x20, 0x30(a0)
-                     sd x21, 0x38(a0)
-                     sd x22, 0x40(a0)
-                     sd x23, 0x48(a0)
-                     sd x24, 0x50(a0)
-                     sd x25, 0x58(a0)
-                     sd x26, 0x60(a0)
-                     sd x27, 0x68(a0)
-                     # Save return address on stack
-                     # It is important to change the stack
-                     # pointer after the previous instructions
-                     # Otherwise the wrong sp is saved (x2 == sp)
-                     addi sp, sp, -0x08
-                     sd ra, 0x00(sp)
-                     jalr a2
-                     ld ra, 0x00(sp)
-                     addi sp, sp, 0x08
-                     ret
-                    "
-            )
-        }
     }
 }
 
@@ -382,9 +153,7 @@ pub fn print() {
 const MAX_STACK_SCAN_SLOTS: usize = 512;
 
 fn scan_stack_for_return_addresses(sp: usize, counter: &mut u64) {
-    // Per-CPU kernel stacks are mapped at the top of the address space
     let stack_bottom = 0usize.wrapping_sub(KERNEL_STACK_SIZE);
-    // Validate SP is within the per-CPU kernel stack before scanning
     if sp < stack_bottom {
         info!("  SP {sp:#x} outside kernel stack, cannot scan");
         return;
@@ -396,8 +165,7 @@ fn scan_stack_for_return_addresses(sp: usize, counter: &mut u64) {
     let text_range = LinkerInformation::text_range();
     for i in 0..slots_to_scan {
         let slot_addr = sp.wrapping_add(i * size_of::<usize>());
-        // SAFETY: slot_addr is within the per-CPU kernel stack (validated above).
-        let value = unsafe { (slot_addr as *const usize).read() };
+        let value = sys::memory::read_usize_from_address(slot_addr);
         if text_range.contains(&VirtAddr::new(value)) {
             print_uncertain_stacktrace_frame(*counter, value);
             *counter += 1;
@@ -436,8 +204,10 @@ fn print_stacktrace_frame_inner(counter: u64, address: usize, prefix: &str) {
 #[cfg(not(miri))]
 #[cfg(test)]
 mod tests {
-    use crate::debugging::backtrace::{Backtrace, BacktraceNextError, CalleeSavedRegs};
+    #![allow(unsafe_code)]
+    use crate::debugging::backtrace::{Backtrace, BacktraceNextError};
     use alloc::collections::VecDeque;
+    use arch::backtrace::CalleeSavedRegs;
     use core::ffi::c_void;
     use unwinding::abi::{_Unwind_Backtrace, _Unwind_GetIP, UnwindContext, UnwindReasonCode};
 
@@ -483,7 +253,6 @@ mod tests {
                 }
             }
 
-            // Skip some items because they are inside the unwind functions itself
             data.addresses.pop_front();
             data.addresses.pop_front();
             own_addr.pop_front();
